+++ server/app/controllers/productOrderActions.js
@@ -16,9 +16,17 @@ const browseProductOrder = async (request, response, next) => {
 
 const readProductOrder = async (request, response, next) => {
   try {
-    const { id } = request.params;
+    const { product_id, order_id } = request.params;
 
-    const productOrderDuo = await tables.Product_order.readProductOrder(id);
+    if (!product_id || !order_id) {
+      response.status(400).json({ message: "Missing required fields" });
+      return;
+    }
+
+    const productOrderDuo = await tables.Product_order.readProductOrder(
+      product_id,
+      order_id
+    );
 
     if (!productOrderDuo) {
       response.status(404).json({ message: "Product_order duo not found" });
@@ -30,45 +38,23 @@ const readProductOrder = async (request, response, next) => {
   }
 };
 
-const editProductOrder = async (request, response, next) => {
-  const { product_id, order_id } = request.params;
-
-  const productOrderDuo = request.body;
-
-  try {
-    const affectedRows = await tables.Product_order.updateProductOrder(
-      product_id,
-      order_id,
-      productOrderDuo
-    );
-
-    if (!affectedRows) {
-      response.status(404).json({ message: "Product_order duo not found" });
-    } else {
-      response.status(200).json({ message: "Product_order duo updated" });
-    }
-  } catch (error) {
-    next(error);
-  }
-};
-
 const addProductOrder = async (request, response, next) => {
   try {
-    const productCategory = request.body;
+    const productOrderDuo = request.body;
 
-    if (!productCategory.product_id || !productCategory.order_id) {
+    if (!productOrderDuo.product_id || !productOrderDuo.order_id) {
       response.status(400).json({ message: "Missing required fields" });
       return;
     }
 
-    const insertId =
-      await tables.Product_order.createProductOrder(productCategory);
+    const affectedRows =
+      await tables.Product_order.createProductOrder(productOrderDuo);
 
-    if (!insertId) {
+    if (!affectedRows) {
       response.status(404).json({ message: "Product_order duo not found" });
     } else {
       response.status(201).json({
-        insertId,
+        affectedRows,
         message: "Product_order duo added",
       });
     }
@@ -78,9 +64,14 @@ const addProductOrder = async (request, response, next) => {
 };
 
 const destroyProductOrder = async (request, response, next) => {
-  const { product_id, order_id } = request.params;
-
   try {
+    const { product_id, order_id } = request.params;
+
+    if (!product_id || !order_id) {
+      response.status(400).json({ message: "Missing required fields" });
+      return;
+    }
+
     const affectedRows = await tables.Product_order.deleteProductOrder(
       product_id,
       order_id
@@ -99,7 +90,6 @@ const destroyProductOrder = async (request, response, next) => {
 module.exports = {
   browseProductOrder,
   readProductOrder,
-  editProductOrder,
   addProductOrder,
   destroyProductOrder,
 };
